feat: add sync persistence wrapper
